Move FunctionInjector into astmodel
